fix: 프로필 이미지 삭제 조건에 빈 문자열 체크 추가

+++ src/main/java/com/meongnyangerang/meongnyangerang/service/HostService.java
@@ -198,7 +198,7 @@ public class HostService {
     Host host = hostRepository.findById(hotsId)
         .orElseThrow(() -> new MeongnyangerangException(NOT_EXIST_ACCOUNT));
 
-    if (host.getProfileImageUrl() != null) {
+    if (host.getProfileImageUrl() != null && !host.getProfileImageUrl().isBlank()) {
       imageService.deleteImageAsync(host.getProfileImageUrl());
     }
